src/functionality/requirements.txt
apturl==0.5.2
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.7.0
ceph==1.0.0
cephfs==2.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
gensim==3.8.3
httplib2==0.14.0
idna==2.10
inflect==5.2.0
joblib==1.0.1
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
nltk==3.5
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
prettytable==0.7.2
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
pyfailsafe==0.6.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
regex==2020.11.13
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
scipy==1.6.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
smart-open==4.2.0
snowballstemmer==2.1.0
soupsieve==2.1
systemd-python==234
tqdm==4.58.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
